+++ docs/CHANGELOG.md
@@ -1,5 +1,6 @@
 # Futur release
 ## Breaking changes
+ - **Removed**: `IPrincipal`. Rdd now uses current `ClaimsPrincipal`
  - **Removed**: Unused `UnreachableEntityException`, `RightExpressionsHelper`, `ICombinationsHolder`. Replaced by `Closed/OpenRightExpressionsHelper`. To correctly handle rights, please implement your version, or use external librairy (`Lucca.Core.Rights`).
  - **Modification**: `AddRddRights<TCombinationsHolder, TPrincipal>()` -> `AddRddDefaultRights(RightDefaultMode mode)`
  - **Modification**: `AddRddSerialization<TPrincipal>()` -> `AddRddSerialization()`